Implement solar system loading as SSCDataLoader.
